Rewrite ctb_queue.py to be a functional module
Original CTBQueue class resulting from class-based approach appears
unnecessary for now since the queue (order of party member/enemy turns)
itself doesn't need attributes. The info that could be attributes for a
CTBQueue class are just as easily retrieved from whatever is returned
from a function version and don't need to be stored persistently.
There is also only ever 1 queue active in a given battle or encounter
between the party and enemies, so without need for multiple objects
inheriting the CTBQueue class nor need to track the state, the class
can be dissolved into a function and the script made a module.